// features/unidades/types/unidadeForm.ts
export interface UnidadeFormData {
  empreendimentoId: string;

  numero: string;

  bloco: string;

  torre: string;

  andar: string;

  tipologia: string;

  status: string;

  quartos: string;

  suites: string;

  banheiros: string;

  vagas: string;

  areaPrivativa: string;

  areaTotal: string;

  precoTabela: string;

  precoPromocional: string;

  comissao: string;

  posicaoSolar: string;

  vista: string;

  observacoes: string;
}